+++ src/index.html
@@ -1,5 +1,3 @@
-<!DOCTYPE html>
-<html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -8,11 +6,14 @@
 </head>
 <body>
     <section>
-            <h4 class="todo-item">Today's To Do List  <i class="fas fa-sync" alt="Refresh"></i></h4>
-            <div class="todo-item add-todo"><input type="text" id="todo-input" placeholder="Add a new todo...">
-                <i class="fa-solid fa-plus-circle todo-add-btn"></i></div>
-                <div class="tasks"></div>
-                <p class="todo-item last-para">Clear all completed</p>
+        <h4 class="todo-item">Today's To Do List <i class="fas fa-sync" alt="Refresh"></i></h4>
+        <div class="todo-item add-todo">
+            <input type="text" id="todo-input" placeholder="Add a new todo...">
+            <i class="fa-solid fa-plus-circle todo-add-btn"></i>
+        </div>
+        <div class="tasks"></div>
+        <p class="todo-item last-para clear-completed-btn">Clear all completed</p>
     </section>
+
 </body>
 </html>
