+++ src/style.css
@@ -32,14 +32,7 @@ input {
   border: none;
 }
 
-/* style.css */
-
 /* Add some basic styles to the todo items and the edit/delete buttons */
-
-/* style.css */
-
-/* Add some basic styles to the todo items and the edit/delete buttons */
-
 .todo-item label {
   flex: 1;
   margin-left: 8px;
@@ -62,5 +55,3 @@ input {
 .todo-item:hover .fa-ellipsis-vertical {
   display: none;
 }
-
-/* ... Other styles ... */
